Refactor: Implement ProfileProvider for user profile management; integrate profile loading and error handling in MobileHeader and Sidebar components
