+++ components/auth-provider.tsx
@@ -1,10 +1,11 @@
 "use client"
 
 import type React from "react"
-import { createContext, useEffect, useState } from "react"
+import { createContext, ReactNode, useContext, useEffect, useState } from "react"
 import { onAuthStateChanged } from "firebase/auth"
 import { doc, getDoc } from "firebase/firestore"
 import { auth, db } from "@/lib/firebase"
+import { profileService, UserProfile } from "@/lib/firebase-admin"
 
 interface UserData {
   uid: string
@@ -17,6 +18,13 @@ interface AuthContextType {
   user: UserData | null
   loading: boolean
   userRole?: string
+}
+
+type ProfileContextType = {
+  profile: UserProfile | null
+  loading: boolean
+  error: string | null
+  reloadProfile: () => void
 }
 
 export const AuthContext = createContext<AuthContextType>({
@@ -79,5 +87,61 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
     <AuthContext.Provider value={{ user, loading, userRole }}>
       {children}
     </AuthContext.Provider>
+  )
+}
+
+
+
+
+type Props = {
+  user: any
+  userRole: string
+  children: ReactNode
+}
+
+export const ProfileContext = createContext<ProfileContextType | undefined>(undefined)
+export const ProfileProvider = ({ user, userRole, children }: Props) => {
+  const [profile, setProfile] = useState<UserProfile | null>(null)
+  const [loading, setLoading] = useState<boolean>(true)
+  const [error, setError] = useState<string | null>(null)
+
+  const loadProfile = async () => {
+    if (!user) return
+    setLoading(true)
+    setError(null)
+
+    try {
+      let userProfile = await profileService.getUserProfile(user.uid)
+
+      if (!userProfile) {
+        const newProfile: UserProfile = {
+          id: user.uid,
+          name: user.name || user.email?.split("@")[0] || "User",
+          email: user.email || "",
+          role: userRole as "student" | "teacher" | "admin",
+          joined_at: new Date() as any,
+        }
+
+        await profileService.createUserProfile(newProfile)
+        userProfile = newProfile
+      }
+
+      setProfile(userProfile)
+    } catch (err) {
+      console.error("Error loading profile:", err)
+      setError("Failed to load profile. Please try again.")
+    } finally {
+      setLoading(false)
+    }
+  }
+
+  useEffect(() => {
+    loadProfile()
+  }, [user, userRole])
+
+  return (
+    <ProfileContext.Provider value={{ profile, loading, error, reloadProfile: loadProfile }}>
+      {children}
+    </ProfileContext.Provider>
   )
 }
