+++ hooks/use-auth.ts
@@ -1,12 +1,20 @@
 "use client"
 
 import { useContext } from "react"
-import { AuthContext } from "@/components/auth-provider"
+import { AuthContext, ProfileContext } from "@/components/auth-provider"
 
 export const useAuth = () => {
   const context = useContext(AuthContext)
   if (!context) {
     throw new Error("useAuth must be used within an AuthProvider")
+  }
+  return context
+}
+
+export const useProfile = () => {
+  const context = useContext(ProfileContext)
+  if (!context) {
+    throw new Error("useProfile must be used within a ProfileProvider")
   }
   return context
 }
